flask: Converted /tasks routes to proper blueprint
